Move oauth_state_middleware out of the middleware directory
The new autoloader either wants the class to be scoped with Middleware::OauthStateMiddleware or not be in the middleware directory. I opted to just move it out of that directory.
